QAallevents: fill QA hists regardless of prim vtx (#695)
There is some interest in examining QA histograms for occasions when
there is no primary vertex, including occasions when tracking is
excluded from production (e.g. for calorimeter interests). This
introduces a chain option `QAallevents` to enable bypassing the cut on
primary vertex, and it is independent of the `QAalltrigs` option which
is a bypass of the exclusion based on physics triggers.