Add EPDX to RICH scalers (#704)
The EPD coincidence rate began filling element `rs14` of the RICH scaler
struct in the DAQ stream and online databases at the end of March 2018.
This needs to be propagated into the `trigDetSums` table in the offline
database ([now
done](https://www.star.bnl.gov/cgi-bin/protected/viewvc.cgi/cvsroot/online/DataBase/dbcron/macros/FillTD.C?r1=1.2&r2=1.3)),
and when reading the DAQ structures (this PR). As was previously done
for the no-killer ZDCs, this makes use of a no-longer needed scaler in
the trigDetSums struct (TOFp in this case), while retaining backward
compatibility with older data in which that now-defunct scaler was
actually used.